Python GUI helper: Initial commit of QT-based "UI control" project. This is a UI visualisation of the Fpga registers, with some basic control option.
Code is a bit rough, a few register bits can be controlled by clicking,
config files can be listed and loaded but chip not configured.
TBD what this GUI would need on top, but maybe a simple Readout to
CSV-like method will be helpful to quicklu configure and fetch some
data, especially when testing a new chip or trying to learn about the
system.